use semaphore and unbound coroutine pattern to handle receive connection
semaphore is the same thing in golang WaitGroup

+++ study_group/predis/src/models/configuration.rs
@@ -1,17 +1,17 @@
 pub struct Configuration {
     pub port: i32,
-    pub workers: i32,
+    pub workers: usize,
 }
 
 impl Configuration {
     pub fn new() -> Self {
         const DEFAULT_PORT: i32 = 6379;
-        const DEFAULT_WORKERS: i32 = 1;
+        const DEFAULT_WORKERS: usize = 1;
         let port = std::env::var("PORT").unwrap_or(DEFAULT_PORT.to_string());
         let workers = std::env::var("WORKERS").unwrap_or(DEFAULT_WORKERS.to_string());
         Configuration {
             port: port.parse::<i32>().unwrap_or(DEFAULT_PORT),
-            workers: workers.parse::<i32>().unwrap_or(DEFAULT_WORKERS),
+            workers: workers.parse::<usize>().unwrap_or(DEFAULT_WORKERS),
         }
     }
 }
